backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch==2.0.1
transformers==4.30.2
python-dotenv==0.19.0
google-generativeai==0.3.0
pydantic==1.10.12